pyproject.toml
[project]
name = "asl-web-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "crawl4ai>=0.8.0",
    "dotenv>=0.9.9",
    "ipykernel>=7.2.0",
    "langchain>=1.2.12",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.19",
    "langchain-mistralai>=1.1.2",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.2",
    "nest-asyncio>=1.6.0",
    "playwright>=1.58.0",
    "pydantic>=2.12.5",
    "pymongo>=4.16.0",
    "requests>=2.32.5",
    "ruff>=0.15.6",
    "streamlit>=1.55.0",
    "tavily>=1.1.0",
    "ultimate-sitemap-parser>=1.8.0",
    "validators>=0.35.0",
]

[tool.ruff]
line-length = 88
target-version = "py312"
extend-include = ["*.ipynb"]

[tool.ruff.lint]
select = ["E","F","I","B","UP"]
ignore = ["E501"]
unfixable = ["F401"]

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
